feat : 주문 생성 (동시성 미적용)

+++ src/main/java/com/example/jjangushrine/domain/order/dto/response/OrderItemRes.java
@@ -2,7 +2,7 @@ package com.example.jjangushrine.domain.order.dto.response;
 
 public record OrderItemRes(
         String productName,
-        Short quantity,
+        int quantity,
         int productPrice
 ) {
 }
